+++ utilities/options/options_util.cc
@@ -20,7 +20,8 @@ Status LoadOptionsFromFile(const std::string& file_name, Env* env,
                            std::shared_ptr<Cache>* cache) {
   RocksDBOptionsParser parser;
   LegacyFileSystemWrapper fs(env);
-  Status s = parser.Parse(file_name, &fs, ignore_unknown_options);
+  Status s = parser.Parse(file_name, &fs, ignore_unknown_options,
+                          0 /* file_readahead_size */);
   if (!s.ok()) {
     return s;
   }